chore(collator): split collation manager on several modules
